feat: ScheduleTemplateService 예외 처리 추가

+++ src/main/java/com/burntoburn/easyshift/service/templates/imp/ScheduleTemplateServiceImpl.java
@@ -5,6 +5,7 @@ import com.burntoburn.easyshift.dto.template.req.ScheduleTemplateRequest;
 import com.burntoburn.easyshift.dto.template.ScheduleTemplateWithShiftsResponse;
 import com.burntoburn.easyshift.entity.store.Store;
 import com.burntoburn.easyshift.entity.templates.ScheduleTemplate;
+import com.burntoburn.easyshift.exception.template.TemplateException;
 import com.burntoburn.easyshift.repository.schedule.ScheduleTemplateRepository;
 import com.burntoburn.easyshift.repository.store.StoreRepository;
 import com.burntoburn.easyshift.service.templates.ScheduleTemplateFactory;
@@ -28,7 +29,7 @@ public class ScheduleTemplateServiceImpl implements ScheduleTemplateService {
     public ScheduleTemplateResponse createScheduleTemplate(Long storeId, ScheduleTemplateRequest request) {
         // Store 조회
         Store store = storeRepository.findById(storeId)
-                .orElseThrow(() -> new NoSuchElementException("Store not found with id: " + storeId));
+                .orElseThrow(()->new NoSuchElementException("not found store"));
 
         // ScheduleTemplate 생성
         ScheduleTemplate scheduleTemplate = scheduleTemplateFactory.createScheduleTemplate(store, request);
@@ -42,13 +43,27 @@ public class ScheduleTemplateServiceImpl implements ScheduleTemplateService {
     @Transactional(readOnly = true)
     @Override
     public ScheduleTemplateWithShiftsResponse getAllScheduleTemplatesByStore(Long storeId) {
+        boolean exists = storeRepository.existsById(storeId);
+        if (!exists){
+            throw new NoSuchElementException("not found store");
+        }
+
         List<ScheduleTemplate> scheduleTemplateList = scheduleTemplateRepository.findAllByStoreId(storeId);
+
+        // 찾는 스케줄이 비어 있을 때 예외 발생
+        if(scheduleTemplateList.isEmpty()){
+            throw TemplateException.scheduleTemplateNotFound();
+        }
 
         return ScheduleTemplateWithShiftsResponse.fromEntities(scheduleTemplateList);
     }
 
     @Override
     public void deleteScheduleTemplate(Long scheduleTemplateId) {
+        boolean exists = scheduleTemplateRepository.existsById(scheduleTemplateId);
+        if (!exists){
+            throw TemplateException.scheduleTemplateNotFound();
+        }
         scheduleTemplateRepository.deleteById(scheduleTemplateId);
     }
 }
